pyproject.toml
[project]
name = "dat1-cli"
version = "0.0.1"
license = { text = "MIT" }

dependencies = [
    "typer >= 0.9.0",
    "inquirer >= 3.1.3",
    "boto3 >= 1.26.146",
    "pyyaml >= 6.0.2",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.scripts]
dat1 = "dat1.app:app"
